+++ mutmut/__main__.py
@@ -1,12 +1,11 @@
 # coding=utf-8
 
-from __future__ import print_function
 
-import itertools
+import argparse
 import os
 import sys
+from configparser import ConfigParser, NoOptionError, NoSectionError
 from datetime import datetime
-from difflib import unified_diff
 from functools import wraps
 from io import open
 from os.path import isdir, exists
@@ -15,27 +14,14 @@ from subprocess import Popen
 from threading import Thread
 from time import sleep
 
-import click
 from glob2 import glob
 
-from mutmut.cache import register_mutants, update_mutant_status, print_result_cache, cached_mutation_status, mutation_id_from_pk, filename_and_mutation_id_from_pk, cached_test_time, set_cached_test_time
-from . import mutate_file, Context, list_mutations, __version__, BAD_TIMEOUT, OK_SUSPICIOUS, BAD_SURVIVED, OK_KILLED, UNTESTED, mutate
+from mutmut.cache import register_mutants, update_mutant_status, \
+    print_result_cache, cached_mutation_status, \
+    filename_and_mutation_id_from_pk, cached_test_time, set_cached_test_time
+from mutmut.mutators import mutate_file, Context, list_mutations, BAD_TIMEOUT, \
+    OK_SUSPICIOUS, BAD_SURVIVED, OK_KILLED, UNTESTED
 from .cache import hash_of_tests
-
-spinner = itertools.cycle('⠋⠙⠹⠸⠼⠴⠦⠧⠇⠏')
-
-if sys.version_info < (3, 0):   # pragma: no cover (python 2 specific)
-    # noinspection PyCompatibility,PyUnresolvedReferences
-    from ConfigParser import ConfigParser, NoOptionError, NoSectionError
-    # This little hack is needed to get the click tester working on python 2.7
-    orig_print = print
-
-    def print(x='', **kwargs):
-        orig_print(x.encode('utf8'), **kwargs)
-
-else:
-    # noinspection PyUnresolvedReferences,PyCompatibility
-    from configparser import ConfigParser, NoOptionError, NoSectionError
 
 
 # decorator
@@ -58,31 +44,8 @@ def config_from_setup_cfg(**defaults):
             f(*args, **kwargs)
 
         return wrapper
+
     return decorator
-
-
-# this function is stolen and modified from tqdm
-def status_printer():
-    """
-    Manage the printing and in-place updating of a line of characters.
-    Note that if the string is longer than a line, then in-place
-    updating may not work (it will print a new line at each refresh).
-    """
-    last_len = [0]
-
-    def p(s):
-        s = next(spinner) + ' ' + s
-        len_s = len(s)
-        output = '\r' + s + (' ' * max(last_len[0] - len_s, 0))
-        if sys.version_info < (3, 0):  # pragma: no cover (python 2 specific)
-            output = output.encode('utf8')
-        sys.stdout.write(output)
-        sys.stdout.flush()
-        last_len[0] = len_s
-    return p
-
-
-print_status = status_printer()
 
 
 def get_or_guess_paths_to_mutate(paths_to_mutate):
@@ -96,7 +59,7 @@ def get_or_guess_paths_to_mutate(paths_to_mutate):
         elif isdir(this_dir):
             return this_dir
         else:
-            raise ErrorMessage('Could not figure out where the code to mutate is. Please specify it on the command line like "mutmut code_dir" or by adding "paths_to_mutate=code_dir" in setup.cfg under the section [mutmut]')
+            raise Exception('Could not figure out where the code to mutate is')
     else:
         return paths_to_mutate
 
@@ -113,18 +76,14 @@ def do_apply(mutation_pk, dict_synonyms, backup):
         context=context,
     )
     if context.number_of_performed_mutations == 0:
-        raise ErrorMessage('ERROR: no mutations performed. Are you sure the index is not too big?')
-
-
-class ErrorMessage(Exception):
-    pass
-
-
-null_out = open(os.devnull, 'w')
+        raise Exception(
+            'No mutations performed. Are you sure the index is not too big?')
 
 
 class Config(object):
-    def __init__(self, swallow_output, test_command, exclude_callback, baseline_time_elapsed, backup, dict_synonyms, total, using_testmon, cache_only, tests_dirs, hash_of_tests):
+    def __init__(self, swallow_output, test_command, exclude_callback,
+                 baseline_time_elapsed, backup, dict_synonyms, total,
+                 using_testmon, cache_only, tests_dirs, hash_of_tests):
         self.swallow_output = swallow_output
         self.test_command = test_command
         self.exclude_callback = exclude_callback
@@ -144,31 +103,50 @@ class Config(object):
         self.suspicious_mutants = 0
 
     def print_progress(self):
-        print_status('%s/%s  🎉 %s  ⏰ %s  🤔 %s  🙁 %s' % (self.progress, self.total, self.killed_mutants, self.surviving_mutants_timeout, self.suspicious_mutants, self.surviving_mutants))
+        print('%s/%s  🎉 %s  ⏰ %s  🤔 %s  🙁 %s' % (
+            self.progress, self.total, self.killed_mutants,
+            self.surviving_mutants_timeout, self.suspicious_mutants,
+            self.surviving_mutants))
 
 
 DEFAULT_TESTS_DIR = 'tests/:test/'
 
 
-@click.command(context_settings=dict(help_option_names=['-h', '--help']))
-@click.argument('command', nargs=1, required=False)
-@click.argument('argument', nargs=1, required=False)
-@click.option('--paths-to-mutate', type=click.STRING)
-@click.option('--backup/--no-backup', default=False)
-@click.option('--runner')
-@click.option('--use-coverage', is_flag=True, default=False)
-@click.option('--tests-dir')
-@click.option('-s', help='turn off output capture', is_flag=True)
-@click.option('--dict-synonyms')
-@click.option('--cache-only', is_flag=True, default=False)
-@click.option('--version', is_flag=True, default=False)
-@config_from_setup_cfg(
-    dict_synonyms='',
-    runner='python -m pytest -x',
-    tests_dir=DEFAULT_TESTS_DIR,
-)
-def main(command, argument, paths_to_mutate, backup, runner, tests_dir, s, use_coverage, dict_synonyms, cache_only, version):
-    """
+def get_argparser() -> argparse.ArgumentParser:
+    """get the main arguement parser for mutmut"""
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--use-coverage", action="store_true",
+                        dest="use_coverage")
+    parser.add_argument("-t", "--tests-dir")
+    parser.add_argument("--paths-to-mutate", default=".", dest="mutate_paths")
+    parser.add_argument("--runner", default="pytest")
+    parser.add_argument("--results", action="store_true")
+    parser.add_argument("--backup", action="store_true")
+    parser.add_argument("--apply")
+    parser.add_argument("--tests-dir", dest="tests_dir")
+    parser.add_argument("-s", action="store_true", dest="output_capture",
+                        help="turn off output capture")
+    parser.add_argument("--cache-only", action="store_true", dest="cache_only")
+    return parser
+
+
+# @click.command(context_settings=dict(help_option_names=['-h', '--help']))
+# @click.argument('command', nargs=1, required=False)
+# @click.argument('argument', nargs=1, required=False)
+# @click.option('--paths-to-mutate', type=click.STRING)
+# @click.option('--backup/--no-backup', default=False)
+# @click.option('--use-coverage', is_flag=True, default=False)
+# @click.option('--tests-dir')
+# @click.option('-s', help='turn off output capture', is_flag=True)
+# @click.option('--dict-synonyms')
+# @click.option('--cache-only', is_flag=True, default=False)
+# @config_from_setup_cfg(
+#     dict_synonyms='',
+#     runner='python -m pytest -x',
+#     tests_dir=DEFAULT_TESTS_DIR,
+# )
+def main():
+    """main entrypoint for mutmut
 commands:\n
     run [mutation id]\n
         Runs mutmut. You probably want to start with just trying this. If you supply a mutation ID mutmut will check just this mutant.\n
@@ -179,76 +157,42 @@ commands:\n
     show [mutation id]\n
         Show a mutation diff.\n
     """
-    if version:
-        print("mutmut version %s" % __version__)
-        return
+    parser = get_argparser()
+    args = parser.parse_args()
+    dict_synonyms = [x.strip() for x in "".split(',')]
+    runner = 'python -m pytest -x'
+    if args.use_coverage and not exists('.coverage'):
+        raise FileExistsError(
+            'No .coverage file found. You must generate a coverage file to use this feature.')
 
-    valid_commands = ['run', 'results', 'apply', 'show']
-    if command not in valid_commands:
-        print('%s is not a valid command, must be one of %s' % (command, ', '.join(valid_commands)))
-        return
-
-    if command == 'results' and argument:
-        print('The %s command takes no arguments' % command)
-        return
-
-    dict_synonyms = [x.strip() for x in dict_synonyms.split(',')]
-
-    if command in ('show', 'diff'):
-        if not argument:
-            print_result_cache()
-            return
-
-        filename, mutation_id = filename_and_mutation_id_from_pk(argument)
-        with open(filename) as f:
-            source = f.read()
-        context = Context(
-            source=source,
-            filename=filename,
-            mutate_id=mutation_id,
-            dict_synonyms=dict_synonyms,
-        )
-        mutated_source, number_of_mutations_performed = mutate(context)
-        if not number_of_mutations_performed:
-            print('No mutation performed')
-            return
-
-        for line in unified_diff(source.split('\n'), mutated_source.split('\n'), fromfile=filename, tofile=filename, lineterm=''):
-            print(line)
-
-        return
-
-    if use_coverage and not exists('.coverage'):
-        print('No .coverage file found. You must generate a coverage file to use this feature.')
-        return
-
-    if command == 'results':
+    if args.results:
         print_result_cache()
+
+    if args.apply:
+        do_apply(args.apply, dict_synonyms, args.backup)
         return
 
-    if command == 'apply':
-        do_apply(argument, dict_synonyms, backup)
-        return
-
-    paths_to_mutate = get_or_guess_paths_to_mutate(paths_to_mutate)
+    paths_to_mutate = get_or_guess_paths_to_mutate(args.mutate_paths)
 
     if not isinstance(paths_to_mutate, (list, tuple)):
         paths_to_mutate = [x.strip() for x in paths_to_mutate.split(',')]
 
     if not paths_to_mutate:
-        raise ErrorMessage('You must specify a list of paths to mutate. Either as a command line argument, or by setting paths_to_mutate under the section [mutmut] in setup.cfg')
+        raise Exception(
+            'You must specify a list of paths to mutate. Either as a command line argument, or by setting paths_to_mutate under the section [mutmut] in setup.cfg')
 
     tests_dirs = []
-    for p in tests_dir.split(':'):
+    for p in args.tests_dir.split(':'):
         tests_dirs.extend(glob(p, recursive=True))
 
     for p in paths_to_mutate:
-        for pt in tests_dir.split(':'):
+        for pt in args.tests_dir.split(':'):
             tests_dirs.extend(glob(p + '/**/' + pt, recursive=True))
-    del tests_dir
 
-    os.environ['PYTHONDONTWRITEBYTECODE'] = '1'  # stop python from creating .pyc files
+    # stop python from creating .pyc files
+    os.environ['PYTHONDONTWRITEBYTECODE'] = '1'
 
+    # TODO:
     using_testmon = '--testmon' in runner
 
     print("""
@@ -270,68 +214,68 @@ Legend for output:
 🙁 Survived. This means your tests needs to be expanded. 
 """)
 
-    try:
-        baseline_time_elapsed = time_test_suite(swallow_output=not s, test_command=runner, using_testmon=using_testmon)
+    baseline_time_elapsed = time_test_suite(
+        swallow_output=not args.output_capture,
+        test_command=runner,
+        using_testmon=using_testmon)
 
-        if using_testmon:
-            copy('.testmondata', '.testmondata-initial')
+    if using_testmon:
+        copy('.testmondata', '.testmondata-initial')
 
-        if not use_coverage:
-            def _exclude(context):
-                return False
-        else:
-            covered_lines_by_filename = {}
-            coverage_data = read_coverage_data(use_coverage)
+    if not args.use_coverage:
+        def _exclude(context):
+            return False
+    else:
+        covered_lines_by_filename = {}
+        coverage_data = read_coverage_data(args.use_coverage)
 
-            def _exclude(context):
-                try:
-                    covered_lines = covered_lines_by_filename[context.filename]
-                except KeyError:
-                    covered_lines = coverage_data.lines(os.path.abspath(context.filename))
-                    covered_lines_by_filename[context.filename] = covered_lines
+        def _exclude(context):
+            try:
+                covered_lines = covered_lines_by_filename[context.filename]
+            except KeyError:
+                covered_lines = coverage_data.lines(
+                    os.path.abspath(context.filename))
+                covered_lines_by_filename[context.filename] = covered_lines
 
-                if covered_lines is None:
-                    return True
-                current_line = context.current_line_index + 1
-                if current_line not in covered_lines:
-                    return True
-                return False
+            if covered_lines is None:
+                return True
+            current_line = context.current_line_index + 1
+            if current_line not in covered_lines:
+                return True
+            return False
 
-        assert command == 'run'
+    mutations_by_file = {}
 
-        mutations_by_file = {}
+    # TODO
+    argument = None
 
-        if argument is None:
-            for path in paths_to_mutate:
-                for filename in python_source_files(path, tests_dirs):
-                    add_mutations_by_file(mutations_by_file, filename, _exclude, dict_synonyms)
-        else:
-            filename, mutation_id = filename_and_mutation_id_from_pk(int(argument))
-            mutations_by_file[filename] = [mutation_id]
+    if argument is None:
+        for path in paths_to_mutate:
+            for filename in python_source_files(path, tests_dirs):
+                add_mutations_by_file(mutations_by_file, filename, _exclude,
+                                      dict_synonyms)
+    else:
+        filename, mutation_id = filename_and_mutation_id_from_pk(int(argument))
+        mutations_by_file[filename] = [mutation_id]
 
-        total = sum(len(mutations) for mutations in mutations_by_file.values())
+    total = sum(len(mutations) for mutations in mutations_by_file.values())
 
-        print()
-        print('2. Checking mutants')
-        config = Config(
-            swallow_output=not s,
-            test_command=runner,
-            exclude_callback=_exclude,
-            baseline_time_elapsed=baseline_time_elapsed,
-            backup=backup,
-            dict_synonyms=dict_synonyms,
-            total=total,
-            using_testmon=using_testmon,
-            cache_only=cache_only,
-            tests_dirs=tests_dirs,
-            hash_of_tests=hash_of_tests(tests_dirs),
-        )
+    print('2. Checking mutants')
+    config = Config(
+        swallow_output=not args.output_capture,
+        test_command=runner,
+        exclude_callback=_exclude,
+        baseline_time_elapsed=baseline_time_elapsed,
+        backup=args.backup,
+        dict_synonyms=dict_synonyms,
+        total=total,
+        using_testmon=using_testmon,
+        cache_only=args.cache_only,
+        tests_dirs=tests_dirs,
+        hash_of_tests=hash_of_tests(tests_dirs),
+    )
 
-        run_mutation_tests(config=config, mutations_by_file=mutations_by_file)
-
-        print()
-    except ErrorMessage as e:
-        print('\nERROR %s' % e)
+    run_mutation_tests(config=config, mutations_by_file=mutations_by_file)
 
 
 def popen_streaming_output(cmd, callback, timeout=None):
@@ -365,7 +309,8 @@ def popen_streaming_output(cmd, callback, timeout=None):
 
     while p.returncode is None:
         try:
-            line = stdout.readline()[:-1]  # -1 to remove the newline at the end
+            line = stdout.readline()[
+                   :-1]  # -1 to remove the newline at the end
             callback(line)
         except OSError:
             # This seems to happen on some platforms, including TravisCI. It seems like
@@ -389,7 +334,8 @@ def tests_pass(config):
             print(line)
         config.print_progress()
 
-    returncode = popen_streaming_output(config.test_command, feedback, timeout=config.baseline_time_elapsed * 10)
+    returncode = popen_streaming_output(config.test_command, feedback,
+                                        timeout=config.baseline_time_elapsed * 10)
     return returncode == 0 or (config.using_testmon and returncode == 5)
 
 
@@ -402,7 +348,8 @@ def run_mutation(config, filename, mutation_id):
         config=config,
     )
 
-    cached_status = cached_mutation_status(filename, mutation_id, config.hash_of_tests)
+    cached_status = cached_mutation_status(filename, mutation_id,
+                                           config.hash_of_tests)
     if cached_status == BAD_SURVIVED:
         config.surviving_mutants += 1
     elif cached_status == BAD_TIMEOUT:
@@ -450,14 +397,14 @@ def run_mutation(config, filename, mutation_id):
 def run_mutation_tests_for_file(config, file_to_mutate, mutations):
     for mutation_id in mutations:
         status = run_mutation(config, file_to_mutate, mutation_id)
-        update_mutant_status(file_to_mutate, mutation_id, status, config.hash_of_tests)
+        update_mutant_status(file_to_mutate, mutation_id, status,
+                             config.hash_of_tests)
         config.progress += 1
         config.print_progress()
 
 
 def fail_on_cache_only(config):
     if config.cache_only:
-        print_status('')
         print('\rFAILED: changes detected in cache only mode')
         exit(2)
 
@@ -488,7 +435,8 @@ def read_coverage_data(use_coverage):
 def time_test_suite(swallow_output, test_command, using_testmon):
     cached_time = cached_test_time()
     if cached_time is not None:
-        print('1. Using cached time for baseline tests, to run baseline again delete the cache file')
+        print(
+            '1. Using cached time for baseline tests, to run baseline again delete the cache file')
         return cached_time
 
     print('1. Running tests without mutations')
@@ -499,7 +447,7 @@ def time_test_suite(swallow_output, test_command, using_testmon):
     def feedback(line):
         if not swallow_output:
             print(line)
-        print_status('Running...')
+        print('Running...')
         output.append(line)
 
     returncode = popen_streaming_output(test_command, feedback)
@@ -507,7 +455,9 @@ def time_test_suite(swallow_output, test_command, using_testmon):
     if returncode == 0 or (using_testmon and returncode == 5):
         baseline_time_elapsed = (datetime.now() - start_time).total_seconds()
     else:
-        raise ErrorMessage("Tests don't run cleanly without mutations. Test command was: %s\n\nOutput:\n\n%s" % (test_command, '\n'.join(output)))
+        raise ErrorMessage(
+            "Tests don't run cleanly without mutations. Test command was: %s\n\nOutput:\n\n%s" % (
+                test_command, '\n'.join(output)))
 
     print(' Done')
 
@@ -528,7 +478,8 @@ def add_mutations_by_file(mutations_by_file, filename, exclude, dict_synonyms):
         mutations_by_file[filename] = list_mutations(context)
         register_mutants(mutations_by_file)
     except Exception:
-        print('Failed while creating mutations for %s, for line "%s"' % (context.filename, context.current_source_line))
+        print('Failed while creating mutations for %s, for line "%s"' % (
+            context.filename, context.current_source_line))
         raise
 
 
@@ -547,7 +498,8 @@ def coverage_exclude_callback(context, use_coverage, coverage_data):
 def python_source_files(path, tests_dirs):
     if isdir(path):
         for root, dirs, files in os.walk(path):
-            dirs[:] = [d for d in dirs if os.path.join(root, d) not in tests_dirs]
+            dirs[:] = [d for d in dirs if
+                       os.path.join(root, d) not in tests_dirs]
             for filename in files:
                 if filename.endswith('.py'):
                     yield os.path.join(root, filename)
@@ -556,7 +508,4 @@ def python_source_files(path, tests_dirs):
 
 
 if __name__ == '__main__':
-    try:
-        main()
-    except ErrorMessage as main_error:
-        print(str(main_error))
+    sys.exit(main())
