+++ mutmut/cache.py
@@ -2,46 +2,41 @@
 
 import hashlib
 import os
-import sys
 from functools import wraps
 from io import open
 from itertools import groupby
 
-from pony.orm import Database, Required, db_session, Set, Optional, select, PrimaryKey
+from pony.orm import Database, Required, db_session, Set, Optional, select, \
+    PrimaryKey
 
-from mutmut import BAD_TIMEOUT, OK_SUSPICIOUS, BAD_SURVIVED, UNTESTED, OK_KILLED, mutant_statuses
-
-if sys.version_info < (3, 0):   # pragma: no cover (python 2 specific)
-    # noinspection PyUnresolvedReferences
-    text_type = unicode
-else:
-    text_type = str
-
+from mutmut.mutators import BAD_TIMEOUT, OK_SUSPICIOUS, BAD_SURVIVED, UNTESTED, \
+    OK_KILLED
 
 db = Database()
 
 
 class MiscData(db.Entity):
-    key = PrimaryKey(text_type, auto=True)
-    value = Optional(text_type, autostrip=False)
+    key = PrimaryKey(str, auto=True)
+    value = Optional(str, autostrip=False)
 
 
 class SourceFile(db.Entity):
-    filename = Required(text_type, autostrip=False)
+    filename = Required(str, autostrip=False)
     lines = Set('Line')
 
 
 class Line(db.Entity):
     sourcefile = Required(SourceFile)
-    line = Required(text_type, autostrip=False)
+    line = Required(str, autostrip=False)
     mutants = Set('Mutant')
 
 
 class Mutant(db.Entity):
     line = Required(Line)
     index = Required(int)
-    tested_against_hash = Optional(text_type, autostrip=False)
-    status = Required(text_type, autostrip=False)  # really an enum of mutant_statuses
+    tested_against_hash = Optional(str, autostrip=False)
+    status = Required(str,
+                      autostrip=False)  # really an enum of mutant_statuses
 
 
 def init_db(f):
